+++ tests/testthat/test-match.r
@@ -50,6 +50,10 @@ test_that("match returns NA when optional group doesn't match", {
   expect_equal(str_match(c("ab", "a"), "(a)(b)?")[,3], c("b", NA))
 })
 
+test_that("match_all returns NA when option group doesn't match",{
+  expect_equal(str_match_all("a", "(a)(b)?")[[1]][1, ], c("a", "a", NA))
+})
+
 test_that("multiple match works", {
   phones_one <- str_c(phones, collapse = " ")
   multi_match <- str_match_all(phones_one,